Solar/solarapi/requirements.txt
absl-py==0.11.0
align==0.0.11
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
blessings==1.7
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==6.7
colorama==0.4.4
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.3
djangorestframework==3.12.2
ee==0.2
entrypoints==0.3
et-xmlfile==1.0.1
example==0.1.0
ExifRead==2.3.2
fasteners==0.15
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gensim==3.8.3
geographiclib==1.50
geojson==2.5.0
geopy==2.0.0
gmaps==0.9.0
google==3.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googlemaps==4.4.2
gpsphoto==2.2.3
grpcio==1.33.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
image==1.5.33
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
lazy-object-proxy==1.4.3
maps==5.1.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
openpyxl==3.0.5
opt-einsum==3.3.0
osgeo==0.0.0
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
piexif==1.1.3
Pillow==8.0.1
pixellib==0.5.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pylab-sdk==1.3.2
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
smart-open==4.0.1
soupsieve==2.0.1
sqlparse==0.4.1
tb-nightly==1.14.0a20190301
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-cpu @ file:///C:/Users/admin/Downloads/tensorflow_cpu-2.3.0-cp37-cp37m-win_amd64.whl
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019030115
threadpoolctl==2.1.0
tifffile==2020.11.26
toml==0.10.2
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
typed-ast==1.4.1
urllib3==1.26.2
virtualenv==20.2.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xlwt==1.3.0
zipp==3.4.0
